requirements.txt
aiohttp==3.6.2
discord.py==1.3.3
PixivPy-Async==1.2.10
python-dotenv==0.13.0
